fluentd, elasticsearch, kibana 테스트

+++ stop.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 set -e
 
-docker-compose down --remove-orphans
-docker system prune -a -f
+TEMP=$1
+IS_PRUNE=${TEMP:-true}
+
+docker-compose --env-file docker-compose.env down --remove-orphans
+
+if [ $IS_PRUNE = true ]; then
+    docker system prune -a -f
+fi
